% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{create_project}
\alias{create_project}
\title{Create a reproducible BioGeoSyn project directory}
\usage{
create_project(output_dir, overwrite = FALSE)
}
\arguments{
\item{output_dir}{Directory where workflow outputs should be written.}

\item{overwrite}{Logical. If \code{FALSE}, existing files are not removed or
overwritten.}
}
\value{
A list of project paths.
}
\description{
Create a reproducible BioGeoSyn project directory
}
