% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the future iBiogeobears Shiny application}
\usage{
launch_app()
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Launch the future iBiogeobears Shiny application
}
