% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_dispersal_network}
\alias{plot_dispersal_network}
\title{Plot a region-to-region dispersal network}
\usage{
plot_dispersal_network(event_table)
}
\arguments{
\item{event_table}{Data frame containing \code{source_region}, \code{target_region},
and optionally \code{frequency}.}
}
\value{
A ggraph object.
}
\description{
Plot a region-to-region dispersal network
}
