% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_event_through_time}
\alias{plot_event_through_time}
\title{Plot events through time}
\usage{
plot_event_through_time(event_table)
}
\arguments{
\item{event_table}{Data frame containing \code{event_time} and \code{event_type}.}
}
\value{
A ggplot object.
}
\description{
Plot events through time
}
