% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_model_comparison}
\alias{plot_model_comparison}
\title{Plot model comparison results}
\usage{
plot_model_comparison(comparison)
}
\arguments{
\item{comparison}{Model comparison table returned by \code{\link[=compare_models]{compare_models()}}.}
}
\value{
A ggplot object.
}
\description{
Plot model comparison results
}
