% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_root_state_probabilities}
\alias{plot_root_state_probabilities}
\title{Plot root state probabilities}
\usage{
plot_root_state_probabilities(root_state_probabilities, top_n = 8L)
}
\arguments{
\item{root_state_probabilities}{Table returned in
\code{standardized_tables$root_state_probabilities}.}

\item{top_n}{Number of highest-probability states to show per model.}
}
\value{
A ggplot object.
}
\description{
Plot root state probabilities
}
